Refactor viser-ng
