--- packages/viser-ng/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": [ "./tools/babel-preset" ]
-}
