+++ packages/viser-ng/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @module
+ * @description
+ * Entry point for all public APIs of the Angular Module
+ */
+export * from './src/index';
