+++ packages/viser-ng/package.json
@@ -5,57 +5,32 @@
   "repository": "viserjs/viser-ng",
   "license": "MIT",
   "authors": "DT-FE",
-  "files": [
-    "lib",
-    "es",
-    "umd",
-    "src"
-  ],
-  "main": "lib/index.js",
-  "module": "es/index.js",
-  "types": "src/index.ts",
+  "files": ["lib", "es", "umd", "src"],
+  "main": "dist/bundles/index.umd.js",
+  "module": "dist/index.js",
   "scripts": {
-    "build": "node ./tools/build.js",
-    "clean": "git clean -fdX .",
-    "lint": "tslint src",
-    "analyse": "cross-env NODE_ENV=analyse webpack --progress --config webpack.config.js"
-  },
-  "peerDependencies": {
-    "@angular/core": ">4",
-    "@angular/compiler": ">4",
-    "@angular/compiler-cli": ">4"
+    "cleanup": "rimraf dist/bundles dist/src dist/index.d.ts dist/index.metadata.json dist/index.js dist/index.js.map dist/LICENSE dist/README.md",
+    "bundling": "rollup -c",
+    "minify": "uglifyjs dist/bundles/index.umd.js --screw-ie8 --compress --mangle --comments --output dist/bundles/angular-module-starter.umd.min.js",
+    "copy": "copyfiles LICENSE README.md dist",
+    "build": "npm run cleanup && tsc && npm run bundling && npm run minify && npm run copy",
+    "publish-to-npm": "cd dist && npm publish"
   },
   "dependencies": {
-    "viser": "^1.0.3"
+    "@angular/core": "^2.4.0",
+    "reflect-metadata": "^0.1.8",
+    "rxjs": "^5.0.3",
+    "viser": "^1.0.0",
+    "zone.js": "^0.7.2"
   },
   "devDependencies": {
-    "@angular/compiler": "^4.0.1",
-    "@angular/compiler-cli": "^4.0.1",
-    "@angular/core": "^4.0.1",
-    "@types/core-js": "^0.9.43",
-    "@types/node": "7.0.7",
-    "babel-cli": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-preset-env": "^1.6.1",
-    "colors": "^1.1.2",
-    "cross-env": "^5.1.1",
-    "gzip-size": "^4.0.0",
-    "ora": "^1.3.0",
-    "pretty-bytes": "^4.0.2",
-    "rimraf": "^2.6.2",
-    "rxjs": "^5.3.0",
-    "tslint": "^5.5.0",
-    "typescript": "^2.6.1",
-    "webpack": "^3.3.0",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "zone.js": "^0.8.5"
+    "@angular/compiler": "^2.4.0",
+    "@angular/compiler-cli": "^2.4.0",
+    "copyfiles": "^1.0.0",
+    "rimraf": "^2.5.4",
+    "rollup": "^0.37.0",
+    "typescript": "~2.4.2",
+    "uglify-js": "^2.7.5"
   },
-  "keywords": [
-    "g2",
-    "chart",
-    "angular",
-    "datavis"
-  ]
+  "keywords": ["g2", "chart", "angular", "datavis"]
 }
