+++ packages/viser-ng/rollup.config.js
@@ -1,0 +1,10 @@
+export default {
+  entry: 'dist/index.js',
+  dest: 'dist/bundles/angular-module-starter.umd.js',
+  sourceMap: false,
+  format: 'umd',
+  moduleName: 'ng.moduleStarter',
+  globals: {
+    '@angular/core': 'ng.core'
+  }
+}
