+++ packages/viser-ng/src/subChart/Axis.ts
@@ -1,6 +1,6 @@
 import * as Style from './Style';
 import { Component, Input } from '@angular/core';
-import { Chart } from '../component';
+import { Chart } from '../Chart';
 
 type formatterFunc = (val: number) => string | number;
