+++ packages/viser-ng/src/subChart/Coord.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Chart } from '../component';
+import { Chart } from '../Chart';
 
 @Component({
   selector: 'Coord',
