+++ packages/viser-ng/src/subChart/Guide.ts
@@ -1,7 +1,7 @@
 
 import * as Style from './Style';
 import { Component, Input } from '@angular/core';
-import { Chart } from '../component';
+import { Chart } from '../Chart';
 type func = () => void;
 
 interface ILineText {
