+++ packages/viser-ng/src/subChart/Tooltip.ts
@@ -1,6 +1,6 @@
 import * as Style from './Style';
 import { Component, Input } from '@angular/core';
-import { Chart } from '../component';
+import { Chart } from '../Chart';
 
 interface ICrosshairs {
   type?: 'rect' | 'x' | 'y' | 'cross',
