--- packages/viser-ng/tools/babel-preset.js
@@ -1,26 +1,0 @@
-const BABEL_ENV = process.env.BABEL_ENV
-const building = BABEL_ENV != undefined && BABEL_ENV !== 'cjs'
-
-const plugins = []
-
-if (BABEL_ENV === 'umd') {
-  plugins.push('external-helpers');
-}
-
-if (process.env.NODE_ENV === 'production') {
-  plugins.push(
-    'dev-expression',
-  );
-}
-
-module.exports = {
-  presets: [
-    [ 'env', {
-      modules: building ? false : 'commonjs',
-      targets: {
-        browsers: ['last 2 versions']
-      }
-    } ],
-  ],
-  plugins: plugins,
-};
