--- packages/viser-ng/webpack.config.js
@@ -1,65 +1,0 @@
-const path = require('path');
-const webpack = require('webpack');
-const env = process.env.NODE_ENV;
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
-
-let config = {
-  entry: './lib/index',
-
-  output: {
-    filename: './umd/viser-ng.js',
-    library: 'ViserNg',
-    libraryTarget: 'umd'
-  },
-
-  resolve: {
-    extensions: ['.ts', '.tsx', '.js', '.json']
-  },
-
-  externals: {},
-
-  module: {
-    loaders: [
-      {
-        test: /\.tsx?$/,
-        exclude: /node_modules/,
-        loaders: [
-          {
-            loader: 'ts-loader'
-          }
-        ]
-      }
-    ]
-  },
-
-  plugins: [
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify(env)
-    })
-  ]
-};
-
-if (env === 'production') {
-  config.plugins.push(
-    new webpack.optimize.UglifyJsPlugin({
-      compressor: {
-        pure_getters: true,
-        unsafe: true,
-        unsafe_comps: true,
-        warnings: false
-      },
-      output: {
-        comments: false
-      },
-      sourceMap: false
-    })
-  );
-
-  config.output.filename = './umd/viser-ng.min.js';
-}
-
-if (env === 'analyse') {
-  config.plugins.push(new BundleAnalyzerPlugin());
-}
-
-module.exports = config;
